Refactor LoadingSpinner
