+++ web/app/post/[id]/page.tsx
@@ -2,7 +2,7 @@
 
 import BackButton from "@/components/general/BackButton";
 import ErrorMessage from "@/components/general/ErrorMessage";
-import LoadingSpinner from "@/components/LoadingSpinner";
+import LoadingSpinner from "@/components/general/LoadingSpinner";
 import PostActions from "@/components/PostActions";
 import UserProfile from "@/components/user/UserProfile";
 import { Post } from "@/types";
